startDownloadFunction/requirements.txt
google-cloud-firestore
google-cloud-pubsub
google-cloud-bigquery
yt-dlp